refactor: use enums instead of plain ints

+++ uwc-backend/Communications/Mcp/UpdateMcpCurrentLoad.cs
@@ -5,6 +5,6 @@ namespace Communications.Mcp
     public class UpdateMcpCurrentLoad
     {
         [JsonProperty("Id")] public int Id { get; set; }
-        [JsonProperty("CurrentLoad")] public double CurrentLoad { get; set; }
+        [JsonProperty("CurrentLoad")] public float CurrentLoad { get; set; }
     }
 }
